Add files via upload
Based on Example 10, create a program that calculates logical expressions. In the editors for x, y, the user writes values separated by a space. For example, the values of the vector x = ”1 0", the values vector y = ”0 1". And in the formula editor , a logical expression, for example, F = “not (x1 and y1) or (x2 and y2) = 1” and the result is calculated by pressing the button.